Refactor modules to be more usable as libraries.
