+++ blint/lib/analysis.py
@@ -10,7 +10,6 @@ from itertools import islice
 from pathlib import Path
 
 import yaml
-from rich.progress import Progress
 from rich.terminal_theme import MONOKAI
 
 from blint.lib.binary import parse
@@ -315,18 +314,7 @@ def print_reviews_table(reviews, files):
     console.print(table)
 
 
-def json_serializer(obj):
-    """JSON serializer to help serialize problematic types such as bytes"""
-    if isinstance(obj, bytes):
-        try:
-            return obj.decode("utf-8")
-        except UnicodeDecodeError:
-            return ""
-
-    return obj
-
-
-def report(src_dir, reports_dir, findings, reviews, files, fuzzables):
+def report(blint_options, findings, reviews, fuzzables):
     """Generates a report based on the analysis results.
 
     Args:
@@ -338,354 +326,28 @@ def report(src_dir, reports_dir, findings, reviews, files, fuzzables):
         fuzzables: A list of fuzzable methods.
 
     """
+    if not findings and not reviews:
+        LOG.info(f":white_heavy_check_mark: No issues found in {blint_options.src_dir_image}!")
+        return
+    
+    if not os.path.exists(blint_options.reports_dir):
+        os.makedirs(blint_options.reports_dir)
     run_uuid = os.environ.get("SCAN_ID", str(uuid.uuid4()))
     common_metadata = {
         "scan_id": run_uuid,
         "created": f"{datetime.now():%Y-%m-%d %H:%M:%S%z}",
     }
     if findings:
-        print_findings_table(findings, files)
-        export_metadata(reports_dir, {**common_metadata, "findings": findings}, "findings")
+        print_findings_table(findings, blint_options.files)
+        export_metadata(blint_options.reports_dir, {**common_metadata, "findings": findings}, "Findings")
     if reviews:
-        print_reviews_table(reviews, files)
-        export_metadata(reports_dir, {**common_metadata, "reviews": reviews}, "reviews")
+        print_reviews_table(reviews, blint_options.files)
+        export_metadata(blint_options.reports_dir, {**common_metadata, "reviews": reviews}, "Reviews")
     if fuzzables:
-        export_metadata(reports_dir, {**common_metadata, "fuzzables": fuzzables}, "fuzzables")
+        export_metadata(blint_options.reports_dir, {**common_metadata, "fuzzables": fuzzables}, "Fuzzables")
     else:
         LOG.debug("No suggestion available for fuzzing")
-
-    if not findings and not reviews:
-        LOG.info(f":white_heavy_check_mark: No issues found in {src_dir}!")
     # Try console output as html
-    else:
-        html_file = Path(reports_dir) / "blint-output.html"
-        console.save_html(html_file, theme=MONOKAI)
-        LOG.info(f"HTML report written to {html_file}")
-
-
-class AnalysisRunner:
-    """Class to analyze binaries."""
-
-    def __init__(self):
-        self.findings = []
-        self.reviews = []
-        self.fuzzables = []
-        self.progress = Progress(
-            transient=True,
-            redirect_stderr=True,
-            redirect_stdout=True,
-            refresh_per_second=1,
-        )
-        self.task = None
-        self.reviewer = None
-
-    def start(self, files, reports_dir, no_reviews=False, suggest_fuzzables=True):
-        """Starts the analysis process for the given source files.
-
-        This function takes the command-line arguments and the reports
-        directory as input, and starts the analysis process. It iterates over
-        the source files, parses the metadata, checks the security properties,
-        performs symbol reviews, and suggests fuzzable targets if specified.
-
-        Args:
-            files (list): The list of source files to be analyzed.
-            reports_dir (str): The directory where the reports will be stored.
-            no_reviews (bool): Whether to perform reviews or not.
-            suggest_fuzzables (bool): Whether to suggest fuzzable targets or not.
-
-        Returns:
-            tuple: A tuple of the findings, reviews, files, and fuzzables.
-        """
-        with self.progress:
-            self.task = self.progress.add_task(
-                f"[green] BLinting {len(files)} binaries",
-                total=len(files),
-                start=True,
-            )
-            for f in files:
-                self._process_files(f, reports_dir, no_reviews, suggest_fuzzables)
-        return self.findings, self.reviews, self.fuzzables
-
-    def _process_files(self, f, reports_dir, no_reviews, suggest_fuzzables):
-        """Processes the given file and generates findings.
-
-        Args:
-            f (str): The file to be processed.
-            reports_dir (str): The directory where the reports will be stored.
-            no_reviews (bool): Whether to perform reviews or not.
-            suggest_fuzzables (bool): Whether to suggest fuzzable targets or not.
-
-        """
-        self.progress.update(self.task, description=f"Processing [bold]{f}[/bold]")
-        metadata = parse(f)
-        exe_name = metadata.get("name", "")
-        # Store raw metadata
-        export_metadata(reports_dir, metadata, f"{os.path.basename(exe_name)}-metadata")
-        self.progress.update(self.task, description=f"Checking [bold]{f}[/bold] against rules")
-        if finding := run_checks(f, metadata):
-            self.findings += finding
-        # Perform symbol reviews
-        if not no_reviews:
-            self.do_review(exe_name, f, metadata)
-        # Suggest fuzzable targets
-        if suggest_fuzzables and (fuzzdata := run_prefuzz(metadata)):
-            self.fuzzables.append(
-                {
-                    "filename": f,
-                    "exe_name": exe_name,
-                    "methods": fuzzdata,
-                }
-            )
-        self.progress.advance(self.task)
-
-    def do_review(self, exe_name, f, metadata):
-        """Performs a review of the given file."""
-        self.progress.update(self.task, description="Checking methods against review rules")
-        self.reviewer = ReviewRunner()
-        self.reviewer.run_review(metadata)
-        if self.reviewer.results:
-            review = self.reviewer.process_review(f, exe_name)
-            self.reviews += review
-
-
-class ReviewRunner:
-    """Class for running reviews."""
-
-    def __init__(self):
-        self.results = {}
-        self.review_methods_list = []
-        self.review_exe_list = []
-        self.review_symbols_list = []
-        self.review_imports_list = []
-        self.review_entries_list = []
-
-    def run_review(self, metadata):
-        """
-        Runs a review of the given file and metadata.
-
-        This function performs a review of the file and metadata based on the
-        available review methods for the executable type. It collects the
-        results from different review methods, including methods for functions,
-        symbols, imports, and dynamic entries.
-
-        Returns:
-            dict[str, list]: Review results where the keys are the review
-            method IDs and the values are lists of matching results.
-        """
-        if not review_methods_dict:
-            LOG.warning("No review methods loaded!")
-            return {}
-        if not metadata or not (exe_type := metadata.get("exe_type")):
-            return {}
-        self._gen_review_lists(exe_type)
-        # Check if reviews are available for this exe type
-        if (
-            self.review_methods_list
-            or self.review_exe_list
-            or self.review_symbols_list
-            or self.review_imports_list
-            or self.review_entries_list
-        ):
-            return self._review_lists(metadata)
-        return self._review_loader_symbols(metadata)
-
-    def _review_lists(self, metadata):
-        """
-        Reviews lists in the metadata and performs specific actions based on the
-        review type.
-
-        Args:
-            metadata (dict): The metadata to review.
-
-        Returns:
-            dict: The results of the review.
-        """
-        if self.review_methods_list or self.review_exe_list:
-            self._methods_or_exe(metadata)
-        if self.review_symbols_list or self.review_exe_list:
-            self._review_symbols_exe(metadata)
-        if self.review_imports_list:
-            self._review_imports(metadata)
-        if self.review_entries_list:
-            self._review_entries(metadata)
-        self._review_pii(metadata)
-        self._review_loader_symbols(metadata)
-        return self.results
-
-    def _review_imports(self, metadata):
-        """
-        Reviews imports in the metadata.
-
-        Args:
-            metadata (dict): The metadata to review.
-        """
-        imports_list = [f.get("name", "") for f in metadata.get("imports", [])]
-        LOG.debug(f"Reviewing {len(imports_list)} imports")
-        self.run_review_methods_symbols(self.review_imports_list, imports_list)
-
-    def _review_entries(self, metadata):
-        """
-        Reviews entries in the metadata and performs specific actions based on
-        the review type.
-
-        Args:
-            metadata (dict): The metadata to review.
-
-        Returns:
-            dict: The results of the review.
-        """
-        entries_list = [
-            f.get("name", "")
-            for f in metadata.get("dynamic_entries", [])
-            if f.get("tag") == "NEEDED"
-        ]
-        LOG.debug(f"Reviewing {len(entries_list)} dynamic entries")
-        self.run_review_methods_symbols(self.review_entries_list, entries_list)
-
-    def _review_pii(self, metadata):
-        """
-        Reviews pii symbols.
-
-        Args:
-            metadata (dict): The metadata to review.
-
-        Returns:
-            dict: The results of the review.
-        """
-        entries_list = [f.get("name", "") for f in metadata.get("pii_symbols", [])]
-        results = defaultdict(list)
-        for e in entries_list[0:EVIDENCE_LIMIT]:
-            results["PII_READ"].append({"pattern": e, "function": e})
-        self.results |= results
-
-    def _review_loader_symbols(self, metadata):
-        """
-        Reviews loader symbols.
-
-        Args:
-            metadata (dict): The metadata to review.
-
-        Returns:
-            dict: The results of the review.
-        """
-        entries_list = [f.get("name", "") for f in metadata.get("first_stage_symbols", [])]
-        results = defaultdict(list)
-        for e in entries_list[0:EVIDENCE_LIMIT]:
-            results["LOADER_SYMBOLS"].append({"pattern": e, "function": e})
-        self.results |= results
-
-    def _review_symbols_exe(self, metadata):
-        """
-        Reviews symbols in the metadata.
-
-        Args:
-            metadata (dict): The metadata to review.
-        """
-        symbols_list = [f.get("name", "") for f in metadata.get("dynamic_symbols", [])]
-        symbols_list += [f.get("name", "") for f in metadata.get("symtab_symbols", [])]
-        LOG.debug(f"Reviewing {len(symbols_list)} symbols")
-        if self.review_symbols_list:
-            self.run_review_methods_symbols(self.review_symbols_list, symbols_list)
-        if self.review_exe_list:
-            self.run_review_methods_symbols(self.review_exe_list, symbols_list)
-
-    def _methods_or_exe(self, metadata):
-        """
-        Reviews lists in the metadata and performs specific actions based on the
-        review type.
-
-        Args:
-            metadata (dict): The metadata to review.
-        """
-        functions_list = [
-            re.sub(r"[*&()]", "", f.get("name", "")) for f in metadata.get("functions", [])
-        ]
-        if metadata.get("magic", "").startswith("PE"):
-            functions_list += [f.get("name", "") for f in metadata.get("symtab_symbols", [])]
-        # If there are no function but static symbols use that instead
-        if not functions_list and metadata.get("symtab_symbols"):
-            functions_list = [f.get("name", "") for f in metadata.get("symtab_symbols", [])]
-        LOG.debug(f"Reviewing {len(functions_list)} functions")
-        if self.review_methods_list:
-            self.run_review_methods_symbols(self.review_methods_list, functions_list)
-        if self.review_exe_list:
-            self.run_review_methods_symbols(self.review_exe_list, functions_list)
-
-    def _gen_review_lists(self, exe_type):
-        """
-        Generates the review lists based on the given executable type.
-
-        This function takes the executable type as input and generates the
-        review lists for the corresponding type. It retrieves the review lists
-        from the review dictionaries based on the executable type.
-        """
-        self.review_methods_list = review_methods_dict.get(exe_type)
-        self.review_exe_list = review_exe_dict.get(exe_type)
-        self.review_symbols_list = review_symbols_dict.get(exe_type)
-        self.review_imports_list = review_imports_dict.get(exe_type)
-        self.review_entries_list = review_entries_dict.get(exe_type)
-
-    def process_review(self, f, exe_name):
-        """
-        Processes the review results for the given executable and review.
-
-        Returns:
-            list[dict]: The processed review result.
-        """
-        reviews = []
-        if not self.results:
-            return {}
-        for cid, evidence in self.results.items():
-            aresult = {
-                **review_rules_cache.get(cid),
-                "evidence": evidence,
-                "filename": f,
-                "exe_name": exe_name,
-            }
-            del aresult["patterns"]
-            reviews.append(aresult)
-        return reviews
-
-    def run_review_methods_symbols(self, review_list, functions_list):
-        """Runs a review of methods and symbols based on the provided lists.
-
-        This function takes a list of review methods and a list of functions and
-        performs a review to find matches between the patterns specified in the
-        review methods and the functions. It returns a dictionary of results
-        where the keys are the review method IDs and the values are lists of
-        matching results.
-
-        Args:
-            review_list (list[dict]): The review methods or symbols.
-            functions_list (list): A list of functions to be reviewed.
-
-        Returns:
-            dict[str, list]: Method/symbol IDs and their results.
-        """
-        results = defaultdict(list)
-        found_cid = defaultdict(int)
-        found_pattern = defaultdict(int)
-        found_function = {}
-        if not review_list:
-            return
-        for review_methods in review_list:
-            for cid, rule_obj in review_methods.items():
-                if found_cid[cid] > EVIDENCE_LIMIT:
-                    continue
-                patterns = rule_obj.get("patterns")
-                for apattern in patterns:
-                    if found_pattern[apattern] > EVIDENCE_LIMIT or found_cid[cid] > EVIDENCE_LIMIT:
-                        continue
-                    for afun in functions_list:
-                        if apattern.lower() in afun.lower() and not found_function.get(
-                            afun.lower()
-                        ):
-                            result = {
-                                "pattern": apattern,
-                                "function": afun,
-                            }
-                            results[cid].append(result)
-                            found_cid[cid] += 1
-                            found_pattern[apattern] += 1
-                            found_function[afun.lower()] = True
-        self.results |= results
+    html_file = Path(blint_options.reports_dir) / "blint-output.html"
+    console.save_html(html_file, theme=MONOKAI)
+    LOG.info(f"HTML report written to {html_file}")
