+++ blint/lib/android.py
@@ -5,6 +5,9 @@ import sys
 import tempfile
 
 from blint.lib.binary import parse, parse_dex
+from custom_json_diff.lib.utils import file_read
+
+from blint.binary import parse, parse_dex
 from blint.config import SYMBOL_DELIMITER
 from blint.cyclonedx.spec import (
     Component,
@@ -15,6 +18,7 @@ from blint.cyclonedx.spec import (
 )
 from blint.logger import LOG
 from blint.lib.utils import check_command, create_component_evidence, find_files, unzip_unsafe
+
 
 ANDROID_HOME = os.getenv("ANDROID_HOME")
 APKANALYZER_CMD = os.getenv("APKANALYZER_CMD")
