+++ blint/lib/utils.py
@@ -5,11 +5,13 @@ import os
 import re
 import shutil
 import string
+import sys
 import tempfile
 import zipfile
+from dataclasses import field
 from importlib.metadata import distribution
 from pathlib import Path
-from typing import Dict
+from typing import Dict, List
 
 import lief
 from ar import Archive
@@ -19,7 +21,6 @@ from orjson import orjson
 from rich import box
 from rich.table import Table
 
-from blint.analysis import json_serializer
 from blint.config import (
     ignore_directories,
     ignore_files,
@@ -561,6 +562,19 @@ def export_metadata(directory: str, metadata: Dict, mtype: str):
     """
     Exports metadata to file.
     """
+    if not os.path.exists(directory):
+        os.makedirs(directory)
     outfile = str(Path(directory) / f"{mtype.lower()}.json")
     output = orjson.dumps(metadata, default=json_serializer).decode("utf-8", "ignore")
     file_write(outfile, output, success_msg=f"{mtype} written to {outfile}", log=LOG)
+
+
+def json_serializer(obj):
+    """JSON serializer to help serialize problematic types such as bytes"""
+    if isinstance(obj, bytes):
+        try:
+            return obj.decode("utf-8")
+        except UnicodeDecodeError:
+            return ""
+
+    return obj
